cli/workshop/requirements.txt
questionary
datetime
pyyaml
pyfiglet
plumbum
ruamel.yaml
